+++ packages/assets-controllers/src/MultichainAssetsRatesController/index.ts
@@ -6,6 +6,10 @@ export type {
   MultichainAssetsRatesControllerStateChange,
   MultichainAssetsRatesControllerMessenger,
 } from './MultichainAssetsRatesController';
+export type {
+  MultichainAssetsRatesControllerUpdateAssetsRatesAction,
+  MultichainAssetsRatesControllerFetchHistoricalPricesForAssetAction,
+} from './MultichainAssetsRatesController-method-action-types';
 
 export {
   MultichainAssetsRatesController,