+++ packages/notification-services-controller/src/NotificationServicesController/index.ts
@@ -1,7 +1,6 @@
-import Controller from './NotificationServicesController';
+import { NotificationServicesController } from './NotificationServicesController';
 
-const NotificationServicesController = Controller;
-export { Controller };
+export { NotificationServicesController as Controller };
 export default NotificationServicesController;
 export * from './NotificationServicesController';
 export type * as Types from './types';
@@ -13,3 +12,23 @@ export * from './constants';
 export * as Mocks from './mocks';
 export * from '../shared';
 export { isVersionInBounds } from './utils/isVersionInBounds';
+
+export type {
+  NotificationServicesControllerInitAction,
+  NotificationServicesControllerEnablePushNotificationsAction,
+  NotificationServicesControllerDisablePushNotificationsAction,
+  NotificationServicesControllerCheckAccountsPresenceAction,
+  NotificationServicesControllerSetFeatureAnnouncementsEnabledAction,
+  NotificationServicesControllerCreateOnChainTriggersAction,
+  NotificationServicesControllerEnableMetamaskNotificationsAction,
+  NotificationServicesControllerDisableNotificationServicesAction,
+  NotificationServicesControllerDisableAccountsAction,
+  NotificationServicesControllerEnableAccountsAction,
+  NotificationServicesControllerFetchAndUpdateMetamaskNotificationsAction,
+  NotificationServicesControllerGetNotificationsByTypeAction,
+  NotificationServicesControllerDeleteNotificationByIdAction,
+  NotificationServicesControllerDeleteNotificationsByIdAction,
+  NotificationServicesControllerMarkMetamaskNotificationsAsReadAction,
+  NotificationServicesControllerUpdateMetamaskNotificationsListAction,
+  NotificationServicesControllerSendPerpPlaceOrderNotificationAction,
+} from './NotificationServicesController-method-action-types';